fee,notification

+++ app/src/main/java/com/Lyber/ui/fragments/ConfirmInvestmentFragment.kt
@@ -9,6 +9,7 @@ import androidx.navigation.fragment.findNavController
 import com.Lyber.R
 import com.Lyber.databinding.FragmentConfirmInvestmentBinding
 import com.Lyber.ui.portfolio.fragment.PortfolioHomeFragment
+import com.Lyber.viewmodels.PortfolioViewModel
 import com.Lyber.utils.CommonMethods.Companion.addFragment
 import com.Lyber.utils.CommonMethods.Companion.checkInternet
 import com.Lyber.utils.CommonMethods.Companion.clearBackStack
@@ -21,7 +22,6 @@ import com.Lyber.utils.CommonMethods.Companion.gone
 import com.Lyber.utils.CommonMethods.Companion.showProgressDialog
 import com.Lyber.utils.CommonMethods.Companion.visible
 import com.Lyber.utils.Constants
-import com.Lyber.viewmodels.PortfolioViewModel
 import java.util.*
 
 class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>(),
@@ -87,7 +87,7 @@ class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>
                                     if (freq == "none") {
                                         viewModel.oneTimeOrderStrategy(
                                             viewModel.selectedStrategy!!.name,
-                                            viewModel.amount.toFloat().toDouble(),
+                                            viewModel.amount.toDouble(),
                                             it.ownerUuid,
                                         )
                                     } else {
@@ -98,7 +98,7 @@ class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>
                                             viewModel.editEnabledStrategy(
                                                 it.ownerUuid,
                                                 freq,
-                                                viewModel.amount.toFloat().toDouble(),
+                                                viewModel.amount.toDouble(),
                                                 viewModel.selectedStrategy!!.name
                                             )
                                         else
@@ -106,7 +106,7 @@ class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>
                                             viewModel.investStrategy(
                                                 it.ownerUuid,
                                                 freq,
-                                                viewModel.amount.toFloat().toDouble(),
+                                                viewModel.amount.toDouble(),
                                                 viewModel.selectedStrategy!!.name
                                             )
                                     }
@@ -135,7 +135,7 @@ class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>
                 viewModel.amount.decimalPoint()
                     .commaFormattedDecimal(decimal) + " ${Constants.MAIN_ASSET_UPPER}"
             tvValueTotal.text =
-                (viewModel.amount.toFloat() + buyValue).toString()
+                (viewModel.amount.toDouble() + buyValue).toString()
                     .decimalPoint()
                     .commaFormattedDecimal(decimal) + " ${Constants.MAIN_ASSET_UPPER}"
             tvValueLyberFee.text =
@@ -172,8 +172,10 @@ class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>
                         tvValueDepositFee
                     ).gone()
                     //changed fee to 1 percent of the amount
-                    var fee = ((viewModel.amount.toFloat() * 0.5f) / 100.0f)
-                    fee = String.format(Locale.US, "%.${decimal}f", fee).toFloat()
+
+
+                    var fee = ((viewModel.amount.toDouble() * 0.5f) / 100.0f)
+                    fee = String.format(Locale.US, "%.${decimal}f", fee).toDouble()
 
                     tvNestedAmount.text = getString(R.string.invest)
 
@@ -182,14 +184,15 @@ class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>
                     else
                         tvValueFrequency.text = viewModel.selectedFrequency
 
-                    tvNestedAmountValue.text = (viewModel.amount.toFloat() - fee).toString()
-                        .decimalPoint()
-                        .commaFormattedDecimal(decimal) + " ${Constants.MAIN_ASSET_UPPER}"
+
+                    tvNestedAmountValue.text = (viewModel.amount.toDouble() - fee).toString()
+                        .decimalPoint().commaFormattedDecimal(decimal) + " ${Constants.MAIN_ASSET_UPPER}"
+
 
 
 //                    viewModel.amount.decimalPoint().commaFormatted + " ${Constants.MAIN_ASSET_UPPER}"
                     tvValueTotal.text =
-                        (viewModel.amount.toFloat()).toString()
+                        (viewModel.amount.toDouble()).toString()
                             .decimalPoint().commaFormatted + " ${Constants.MAIN_ASSET_UPPER}"
                     tvLyberFee.text = getString(R.string.fee)
                     tvValueLyberFee.text =
@@ -197,7 +200,7 @@ class ConfirmInvestmentFragment : BaseFragment<FragmentConfirmInvestmentBinding>
 
 
                     tvAmount.text =
-                        (viewModel.amount.toFloat()).toString() + " ${Constants.MAIN_ASSET_UPPER}"
+                        (viewModel.amount.toDouble()).toString() + " ${Constants.MAIN_ASSET_UPPER}"
 
                 }
